<Create: New Branch; Fix: Custom Label & Button; Delete: Unused Codes>
