+++ JokeRandomizer/JokeRandomizer/ViewModel/JokesViewModel.swift
@@ -22,8 +22,7 @@ class JokesViewModel {
     enum Output {
         case fetchJokeDidFail(error: Error)
         case fetchJokeSucceed(jokes: [JokesModel])
-//        case toggleButton(isEnabled: Bool)
-//        case toggleLoading(loading: Bool)
+
     }
     
     private let jokeServiceType: JokesServiceType
@@ -39,11 +38,7 @@ class JokesViewModel {
     }
 //MARK: - Func getRandomJokes
     func getRandomJokes() {
-//        output.send(.toggleLoading(loading: true))
-//        output.send(.toggleButton(isEnabled: false))
         jokeServiceType.getRandomJokes().sink { [weak self] completion in
-//            self?.output.send(.toggleLoading(loading: false))
-//            self?.output.send(.toggleButton(isEnabled: true))
             switch completion {
                 case .failure(let errror):
                     self?.output.send(.fetchJokeDidFail(error: errror))
